# Conflicts:
#	README.md
#	extras/Projucer/Builds/MacOSX/Projucer.xcodeproj/project.pbxproj
#	extras/Build/CMake/JUCEUtils.cmake
#	extras/Projucer/Builds/MacOSX/Projucer.xcodeproj/project.pbxproj
#	extras/Projucer/JuceLibraryCode/BinaryData.cpp
#	extras/Projucer/JuceLibraryCode/BinaryData.h
#	extras/Projucer/Source/ProjectSaving/jucer_ProjectExport_CLion.h
#	extras/Projucer/Source/ProjectSaving/jucer_ProjectExport_Xcode.h
#	modules/juce_audio_plugin_client/AU/juce_AU_Wrapper.mm
#	modules/juce_audio_plugin_client/AUResources.r
#	modules/juce_audio_plugin_client/VST3/juce_VST3_Wrapper.cpp
#	modules/juce_audio_plugin_client/juce_audio_plugin_client.h

+++ modules/juce_audio_basics/buffers/juce_FloatVectorOperations.cpp
@@ -353,8 +353,8 @@ namespace FloatVectorHelpers
     union signMask64 { double d; uint64 i; };
 
    #if JUCE_USE_SSE_INTRINSICS || JUCE_USE_ARM_NEON
-    template<int typeSize> struct ModeType    { using Mode = BasicOps32; };
-    template<>             struct ModeType<8> { using Mode = BasicOps64; };
+    template <int typeSize> struct ModeType    { using Mode = BasicOps32; };
+    template <>             struct ModeType<8> { using Mode = BasicOps64; };
 
     template <typename Mode>
     struct MinMax
@@ -873,7 +873,7 @@ void FloatVectorOperations::abs (double* dest, const double* src, int num) noexc
 void JUCE_CALLTYPE FloatVectorOperations::convertFixedToFloat (float* dest, const int* src, float multiplier, int num) noexcept
 {
    #if JUCE_USE_ARM_NEON
-    JUCE_PERFORM_VEC_OP_SRC_DEST (dest[i] = src[i] * multiplier,
+    JUCE_PERFORM_VEC_OP_SRC_DEST (dest[i] = (float) src[i] * multiplier,
                                   vmulq_n_f32 (vcvtq_f32_s32 (vld1q_s32 (src)), multiplier),
                                   JUCE_LOAD_NONE, JUCE_INCREMENT_SRC_DEST, )
    #else
